Integrated ufo-dd, for key backup and key regeneration

+++ src/macbackend.py
@@ -102,6 +102,16 @@ class MacBackend(OSBackend):
                                                   
         return path.join(tmpdir, "UFO.app", "Contents", "MacOS", "UFO")
 
+    def prepare_self_copy(self):
+        self_copied__path = tempfile.mkdtemp(prefix="ufo-self-copied")
+
+        src = path.join(conf.DATA_DIR, "..", "Mac-Intel")
+        logging.debug("Copying " + path.join(src, "UFO.app") + " to " + self_copied__path)
+        self.call([ [ "tar", "-cf", "-", "-C", src, "UFO.app" ],
+                    [ "tar", "xf", "-", "-C", self_copied__path ] ], output = True)[1]
+
+        return path.join(self_copied__path, "UFO.app", "Contents", "MacOS", "UFO")
+
     def get_model(self, dev):
         medianame = utils.grep(self.call(["/usr/sbin/diskutil", "info", dev], output=True)[1], "Media Name:")
         if medianame:
@@ -202,11 +212,11 @@ class MacBackend(OSBackend):
         return blockcount
 
     def find_network_device(self):
-       if conf.NETTYPE == conf.NET_HOST and conf.HOSTNET != "":
-           return conf.NET_HOST, conf.HOSTNET   
+        if conf.NETTYPE == conf.NET_HOST and conf.HOSTNET != "":
+            return conf.NET_HOST, conf.HOSTNET
 
-       return conf.NET_NAT, ""
-  
+        return conf.NET_NAT, ""
+
     # find each mountable partitions of the device
     # and add an entry in fstab to disable automount
     # params : dev_string
@@ -243,49 +253,49 @@ class MacBackend(OSBackend):
             return 0
         return 0
 
-    def check_privileges(self):
-        if os.geteuid() != 0:
-            tries = 0
-            while tries < 3:
-                logging.debug("Asking user password")
-                remember = False
-                password = gui.dialog_password(remember=False)
-                if password == None:
-                    ret = -1
-                    break
+    def execv_as_root(self, executable, cmd):
+        tries = 0
+        while tries < 3:
+            logging.debug("Asking user password")
+            remember = False
+            password = gui.dialog_password(remember=False)
+            if password == None:
+                ret = -1
+                break
 
-                self.call([ "sudo", "-k" ])
-                ret = self.call([ [ "echo", str(password)], 
-                                  [ "sudo", "-S", "touch", sys.executable ] ], log=False)[0]
-                if ret == 0:
-                    if remember:
-                        output = self.call( [ "sudo", "-l" ], output=True)[1]
-                        if not "NOPASSWD: /Volumes/UFO/Mac-Intel/UFO.app/Contents/MacOS/UFO" in output:
-                            sudoline = os.environ["USER"] + " ALL=(ALL) NOPASSWD: /Volumes/UFO/Mac-Intel/UFO.app/Contents/MacOS/UFO"
-                            self.call([ "sudo", "-n", "-s", "echo -e " + sudoline + " >> /etc/sudoers" ])
-                    break
-                else:
-                    gui.dialog_info(title=_("Error"), 
-                                    msg=_("Sorry, couldn't authenticate. Please check your password."),
-                                    error=True)
-                    tries += 1
-
+            self.call([ "sudo", "-k" ])
+            ret = self.call([ [ "echo", str(password)],
+                              [ "sudo", "-S", "touch", sys.executable ] ], log=False)[0]
             if ret == 0:
-                if path.basename(sys.executable) == "python":
-                    cmd = [ path.join(path.dirname(sys.executable), "UFO") ]
-                else:
-                    cmd = [ sys.executable ] + sys.argv
-                cmd += [ "--respawn" ]
-                logging.debug("Environment: " + str(os.environ))
-                logging.debug("Sudoing and execv: " + " ".join([ "/usr/bin/sudo", "-n" ] + cmd))
-                logging.shutdown()
-                if False:
-                    os.execv("/usr/bin/sudo", [ "/usr/bin/sudo", "-n" ] + cmd)
-                    # self.call([ "sudo" ] + cmd, fork=False)
-                else:
-                    self.call([ "sudo" ] + cmd, spawn=True)
+                if remember:
+                    output = self.call( [ "sudo", "-l" ], output=True)[1]
+                    if not "NOPASSWD: /Volumes/UFO/Mac-Intel/UFO.app/Contents/MacOS/UFO" in output:
+                        sudoline = os.environ["USER"] + " ALL=(ALL) NOPASSWD: /Volumes/UFO/Mac-Intel/UFO.app/Contents/MacOS/UFO"
+                        self.call([ "sudo", "-n", "-s", "echo -e " + sudoline + " >> /etc/sudoers" ])
+                break
+            else:
+                gui.dialog_info(title=_("Error"),
+                                msg=_("Sorry, couldn't authenticate. Please check your password."),
+                                error=True)
+                tries += 1
 
-            sys.exit(0)
+        if ret == 0:
+            logging.debug("Environment: " + str(os.environ))
+            logging.debug("Sudoing and execv: " + " ".join([ "/usr/bin/sudo", "-n" ] + cmd))
+            logging.shutdown()
+            if False:
+                os.execv("/usr/bin/sudo", [ "/usr/bin/sudo", "-n" ] + cmd)
+                # self.call([ "sudo" ] + cmd, fork=False)
+            else:
+                self.call([ "sudo" ] + cmd, spawn=True)
+
+    def is_admin(self):
+        return os.geteuid() == 0;
+
+    def umount_device(self, mntpoint):
+        if self.call([ "diskutil", "unmount", mntpoint ]) != 0:
+            return False
+        return True
 
     def get_kernel_modules(self):
         if self.OS_VERSION < "9":
@@ -298,7 +308,7 @@ class MacBackend(OSBackend):
         # loading kernel extentions
         KEXTS = path.join(conf.BIN, self.KEXTS)
         self.tmpdir = tempfile.mkdtemp()
-        
+
         modules = self.get_kernel_modules()
 
         for module in modules:
@@ -320,18 +330,26 @@ class MacBackend(OSBackend):
         self.call([ "killall", "-9", "VBoxSVC" ])
 
     def prepare(self):
-        self.check_privileges()
+        if not self.is_admin():
+            if path.basename(sys.executable) == "python":
+                cmd = [ path.join(path.dirname(sys.executable), "UFO") ]
+            else:
+                cmd = [ sys.executable ] + sys.argv
+            cmd += [ "--respawn" ]
+
+            self.execv_as_root(cmd[0], cmd)
+            exit(1)
 
         if not conf.VBOX_INSTALLED:
             if os.path.islink("/Applications/VirtualBox.app"):
                 os.unlink("/Applications/VirtualBox.app")
-            
+
             # Restore permissions
             # self.call([ "/usr/sbin/chown", "-R", "0:0", conf.APP_PATH ])
             # self.call([ "chmod", "-R", "755", "/Applications/VirtualBox.app/Contents" ])
             # for f in glob.glob("/Applications/VirtualBox.app/Contents/*.*"):
             #     self.call([ "chmod", "-R", "644", f ])
-        
+
             self.load_kexts()
         else:
             self.installed_vbox_error()
